change data type of ContestData.contestPostCount from BigDecimal to int.
git-svn-id: https://coder.topcoder.com/tcs/clients/cronos/components@132305 fb370eea-3af6-4597-97f7-f7400a59c12a

+++ widget_bridge/trunk/src/java/main/com/topcoder/widgets/bridge/AjaxBridgeServlet.java
@@ -1587,7 +1587,7 @@ public class AjaxBridgeServlet extends HttpServlet {
         
         // TCCC-457
         respJSON.setLong("forumId", contest.getForumId());
-        respJSON.setBigDecimal("forumPostCount", contest.getForumPostCount());
+        respJSON.setInt("forumPostCount", contest.getForumPostCount());
 
         return respJSON;
     }
